Created separate teamlineupview module
- Add includeBuild for a root project inside root project. see comment in settings.gradle.kts
- Move HalfFieldDrawable to the new module
