+++ settings.gradle.kts
@@ -5,4 +5,9 @@ plugins {
     id("com.bykov.settings")
 }
 
+dependencyResolutionManagement {
+    // needed for declaring dependencies on other modules using groupId:artifactId notation
+    includeBuild(".")
+}
+
 rootProject.name = "Footea"
