+++ teamlineupview/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    id("com.bykov.android-library")
+}
+
+android {
+    namespace = "com.bykov.footea.teamlineupview"
+}
+
+dependencies {
+
+    // Base
+    implementation("androidx.core:core-ktx")
+}
