+++ teamlineupview/src/main/res/values/colors.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <!-- HalfFieldDrawable -->
+    <color name="field_line">#808080</color>
+</resources>
