+++ teamlineupview/src/main/res/values/dimens.xml
